requirements.txt
asgiref==3.2.7
Authlib==0.14.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
configparser==5.0.0
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.6
django-environ==0.4.5
djangorestframework==3.11.0
idna==2.9
oauth==1.0.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
